refactor: Remove comment property from IMoveNode and MoveVariantNode interfaces, and update related code in RepertoireContext and mock data files to reflect this change. Adjust tests to ensure compatibility with the new structure.
