+++ packages/common/src/types/MoveNode.ts
@@ -4,7 +4,6 @@ export interface IMoveNode {
   id: string;
   move: Move | null;
   children: IMoveNode[];
-  comment?: string;
   variantName?: string;
   circles?: Square[];
   arrows?: Square[][];
