+++ packages/frontend/src/utils/chess/pgn/tests/mocks/repertoire-with-subvariants.mock.ts
@@ -14,10 +14,8 @@ export const testRepertoireWithSubvariantsMock: IMoveNode = {
         san: "e4",
         flags: "b",
         lan: "e2e4",
-        before: "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1",
-        after: "rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq - 0 1",
+        before: "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1",        after: "rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq - 0 1",
       },
-      comment: "",
       children: [
         {
           id: "e7e5",
@@ -31,10 +29,8 @@ export const testRepertoireWithSubvariantsMock: IMoveNode = {
             lan: "e7e5",
             before:
               "rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq - 0 1",
-            after:
-              "rnbqkbnr/pppp1ppp/8/4p3/4P3/8/PPPP1PPP/RNBQKBNR w KQkq - 0 2",
+            after:              "rnbqkbnr/pppp1ppp/8/4p3/4P3/8/PPPP1PPP/RNBQKBNR w KQkq - 0 2",
           },
-          comment: undefined,
           children: [
             {
               id: "g1f3",
@@ -382,10 +378,8 @@ export const testRepertoireWithSubvariantsMock: IMoveNode = {
                                                                 lan: "d1e2",
                                                                 before:
                                                                   "r1bq1rk1/pp3ppp/2nbpn2/2pp4/3P4/2PBPN2/PP1N1PPP/R1BQ1RK1 w - - 5 8",
-                                                                after:
-                                                                  "r1bq1rk1/pp3ppp/2nbpn2/2pp4/3P4/2PBPN2/PP1NQPPP/R1B2RK1 b - - 6 8",
+                                                                after:                                                                  "r1bq1rk1/pp3ppp/2nbpn2/2pp4/3P4/2PBPN2/PP1NQPPP/R1B2RK1 b - - 6 8",
                                                               },
-                                                              comment:undefined,
                                                               variantName:
                                                                 "Sistema Colle",
                                                               children: [],
